[axl_lua][axl_st] minor conversion warning fixes (only happen when compiling against lua 5.3 on 32-bit)

+++ include/axl_lua_LuaState.h
@@ -517,7 +517,7 @@ public:
 	getContext ()
 	{
 		int index = getUpValueIndex ();
-		return getInteger (index);
+		return (intptr_t) getInteger (index);
 	}
 
 	int
